requirements.txt
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
numpy==1.20.3
opencv-contrib-python==4.5.2.52
pdfminer.six==20201018
Pillow==8.2.0
pycparser==2.20
pytesseract==0.3.7
python-magic==0.4.22
sortedcontainers==2.3.0
